cli: add error cause for TCP server's app_loop
If there is an error when running as a TCP server (for instance
a fatal formatting error on the stream), then the CLI will log
the error, close the connection and continue listening for more
TCP connections. This adds the cause of the error to the log,
using Anyhow's Error alternate formatting.

Signed-off-by: Daniel Estévez <daniel@destevez.net>