fix build error with crc-3.2.0
In crc-3.2.0 the crc::Digest struct that we use is no longer Clone.
This is because it has gained an extra type parameter Implementation
that defaults to Table<1>, which does not implement Clone.

This breaks dvb-gse, because the struct Defrag, which includes a
crc::Digest, was deriving Clone.

Since Defrag is a private struct and the fact that it is Clone is
not exposed in the API (it is only included in GSEPacketDefrag, which
is not Clone), we can simply remove the derive Clone from Defrag to
fix this problem.

Signed-off-by: Daniel Estévez <daniel@destevez.net>